deadlock detection: added include/seastar/core/internal/deadlock_utils.hh
This is the interface for helpers used for tracing in deadlock detection.
Each of the helpers is enabled `ifdef SEASTAR_DEADLOCK_DETECTION`.
Otherwise, the helpers are empty functions.

The helpers are used for tracing evolution of runtime graph of tasks,
promises and futures with info about semaphores. Runtime vertex is a
task, a future_base or a promise_base and can be constructed implicitly
from these.

The graph has edges between the vertices. Code that calls `trace_edge`
helper contains comments explaining why the edge is warranted.
Example edges are:
- from promise to it's continuation if `then` was called before promise
  being ready,
- from promise to future and from future to continuation if `then` was
  called on ready future,
- from continuation to promise that represented finishing of continuation.

There is also a concept of speculative edges. Standard edge should
represent pre vertex actually "starting" post vertex.
As an example when promise is completed then it schedules continuation
task. Speculative edge represents any happens-befora relationship, for
example from currently run task to any constructed vertex. That is
 because even if task is created not ready and scheduled later,
the task that crteates it must run first.

Current traced vertex (meaning the vertex that is currently being
executed) can be updated using current_traced_vertex_updater struct.
It works with RAII semantics, updating the current vertex for its
lifetime and reverts when destructed.

Other helpers include tracing semaphore events (waits, signals) and corresponding runtime vertex.

The struct runtime_vertex is needed, so that the implementation of
deadlock detection module doesn't depend on the definions of futures and
promises (as this is irrelevant for this goal, and creates cyclic
dependencies).

The semaphores, being a concrete type, are treated as `void *` (which
is not possible with futures, etc, because address of future_base can be
different that address of future).

Co-authored-by: Michał Radwański <enedil@outlook.com>